+++ src/extension.ts
@@ -8,8 +8,8 @@ import * as path from "path";
 // Your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
   // This register the base command
-  let createBaseStructureCmd = vscode.commands.registerCommand(
-    "mvvmFlutterArchitecture.createBaseStructure",
+  let createProjectStructureCmd = vscode.commands.registerCommand(
+    "mvvmFlutterArchitecture.createProjectStructure",
     async () => {
       const workspaceFolders = vscode.workspace.workspaceFolders;
       if (!workspaceFolders) {
@@ -22,7 +22,7 @@ export function activate(context: vscode.ExtensionContext) {
       const rootPath = workspaceFolders[0].uri.fsPath;
 
       try {
-        await createBaseStructure(rootPath);
+        await createProjectStructure(rootPath);
         vscode.window.showInformationMessage(
           "Base structure created with success!"
         );
@@ -35,8 +35,8 @@ export function activate(context: vscode.ExtensionContext) {
   );
 
   // Command to create a new feature
-  let createFeatureCmd = vscode.commands.registerCommand(
-    "mvvmFlutterArchitecture.createFeature",
+  let createFeatureInPresentationLayerCmd = vscode.commands.registerCommand(
+    "mvvmFlutterArchitecture.createFeatureInPresentationLayer",
     async () => {
       const workspaceFolders = vscode.workspace.workspaceFolders;
       if (!workspaceFolders) {
@@ -58,7 +58,42 @@ export function activate(context: vscode.ExtensionContext) {
       }
 
       try {
-        await createFeatureStructure(rootPath, featureName);
+        await createFeatureInPresentationLayer(rootPath, featureName);
+        vscode.window.showInformationMessage(
+          `Feature "${featureName}" successfully created!`
+        );
+      } catch (error) {
+        vscode.window.showErrorMessage(
+          `An error occurred while creating the feature: ${error}`
+        );
+      }
+    }
+  );
+
+  let createFeatureInAllLayersCmd = vscode.commands.registerCommand(
+    "mvvmFlutterArchitecture.createFeatureInAllLayers",
+    async () => {
+      const workspaceFolders = vscode.workspace.workspaceFolders;
+      if (!workspaceFolders) {
+        vscode.window.showErrorMessage(
+          "Open a workspace folder before running this command."
+        );
+        return;
+      }
+
+      const rootPath = workspaceFolders[0].uri.fsPath;
+
+      // Ask the user to insert the feature name
+      const featureName = await vscode.window.showInputBox({
+        prompt: "Please insert the feature name",
+      });
+      if (!featureName) {
+        vscode.window.showErrorMessage("No feature name inserted.");
+        return;
+      }
+
+      try {
+        await createFeatureInAllLayers(rootPath, featureName);
         vscode.window.showInformationMessage(
           `Feature "${featureName}" successfully created!`
         );
@@ -124,20 +159,20 @@ export function activate(context: vscode.ExtensionContext) {
     }
   );
 
-  context.subscriptions.push(createBaseStructureCmd);
-  context.subscriptions.push(createFeatureCmd);
+  context.subscriptions.push(createProjectStructureCmd);
+  context.subscriptions.push(createFeatureInPresentationLayerCmd);
+  context.subscriptions.push(createFeatureInAllLayersCmd);
   context.subscriptions.push(createBaseConfigurationCmd);
   context.subscriptions.push(createBaseRoutingCmd);
 }
 
-async function createBaseStructure(rootPath: string) {
+async function createProjectStructure(rootPath: string) {
   const folders = [
     "lib/ui/core/ui",
     "lib/ui/core/themes",
     "lib/domain/models",
     "lib/data/repositories",
     "lib/data/services",
-    "lib/data/model",
     "lib/config",
     "lib/utils",
     "lib/routing",
@@ -152,7 +187,7 @@ async function createBaseStructure(rootPath: string) {
   const files = [
     "lib/main_staging.dart",
     "lib/main_development.dart",
-    "lib/main.dart"
+    "lib/main.dart",
   ];
 
   for (const folder of folders) {
@@ -168,7 +203,10 @@ async function createBaseStructure(rootPath: string) {
   }
 }
 
-async function createFeatureStructure(rootPath: string, featureName: string) {
+async function createFeatureInPresentationLayer(
+  rootPath: string,
+  featureName: string
+) {
   const featureFolders = [
     `lib/ui/${featureName}/view_model`,
     `lib/ui/${featureName}/widgets`,
@@ -224,14 +262,18 @@ async function createFeatureStructure(rootPath: string, featureName: string) {
   await fs.promises.writeFile(screenFile, screenContent, { flag: "w" });
 }
 
-async function createBaseConfiguration(rootPath: string) {
+async function createFeatureInAllLayers(rootPath: string, featureName: string) {
+  // The implementation of the addition of the feature in the presentation layer remains the same
+  await createFeatureInPresentationLayer(rootPath, featureName);
+  await createFeatureInDataLayer(rootPath, featureName);
+  await createFeatureInDomainLayer(rootPath, featureName);
+}
 
+async function createBaseConfiguration(rootPath: string) {
   // Adds two files into config folder
   // dependencies.dart handle providers and change notifiers
   // assets.dart handle... assets
-  const configFolders = [
-    "lib/config",
-  ];
+  const configFolders = ["lib/config"];
 
   for (const folder of configFolders) {
     const folderPath = path.join(rootPath, folder);
@@ -263,10 +305,7 @@ List<SingleChildWidget> get providersLocal {
   `;
 
   // Paths to the files
-  const dependenciesFile = path.join(
-    rootPath,
-    `lib/config/dependencies.dart`
-  );
+  const dependenciesFile = path.join(rootPath, `lib/config/dependencies.dart`);
 
   const assetsContent = `
 abstract final class Assets {
@@ -275,21 +314,17 @@ abstract final class Assets {
 }
   `;
 
-  const assetsFile = path.join(
-    rootPath,
-    `lib/config/assets.dart`
-  );
+  const assetsFile = path.join(rootPath, `lib/config/assets.dart`);
 
   // Write the content to the files
-  await fs.promises.writeFile( dependenciesFile, dependenciesContent, { flag: "w" });
-  await fs.promises.writeFile( assetsFile, assetsContent, { flag: "w" });
-
+  await fs.promises.writeFile(dependenciesFile, dependenciesContent, {
+    flag: "w",
+  });
+  await fs.promises.writeFile(assetsFile, assetsContent, { flag: "w" });
 }
 
 async function createBaseRouting(rootPath: string) {
-  const routingFolders = [
-    "lib/routing",
-  ];
+  const routingFolders = ["lib/routing"];
 
   for (const folder of routingFolders) {
     const folderPath = path.join(rootPath, folder);
@@ -300,10 +335,7 @@ async function createBaseRouting(rootPath: string) {
 //! Add your router configuration here
   `;
 
-  const routerFile = path.join(
-    rootPath,
-    `lib/routing/router.dart`
-  );
+  const routerFile = path.join(rootPath, `lib/routing/router.dart`);
 
   const routesContent = `
 abstract final class Routes {
@@ -311,15 +343,10 @@ abstract final class Routes {
   // Add your routes here
 }`;
 
-  const routesFile = path.join(
-    rootPath,
-    `lib/routing/routes.dart`
-  );
+  const routesFile = path.join(rootPath, `lib/routing/routes.dart`);
 
-
-  await fs.promises.writeFile( routerFile, routerContent, { flag: "w" });
-  await fs.promises.writeFile( routesFile, routesContent, { flag: "w" });
-
+  await fs.promises.writeFile(routerFile, routerContent, { flag: "w" });
+  await fs.promises.writeFile(routesFile, routesContent, { flag: "w" });
 }
 
 function capitalizeFeatureName(name: string): string {
@@ -327,6 +354,104 @@ function capitalizeFeatureName(name: string): string {
     .split("_")
     .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
     .join("");
+}
+
+async function createFeatureInDataLayer(rootPath: string, featureName: string) {
+  const featureFolders = [`lib/data/repositories/${featureName}`];
+
+  for (const folder of featureFolders) {
+    const folderPath = path.join(rootPath, folder);
+    await fs.promises.mkdir(folderPath, { recursive: true });
+  }
+
+  // Create the code templates with the feature name
+  const className = capitalizeFeatureName(featureName);
+
+  const repositoryContent = `
+  /// Data source for activities.
+  abstract class ${className}Repository {
+  /// Add the interface methods here
+}`;
+
+  const localRepositoryContent = `
+    import '${featureName}_repository.dart';
+  /// Local implementation of ${className}Repository
+class ${className}RepositoryLocal implements ${className}Repository {
+
+  // In the constructor, you can add the dependencies needed to access the local data
+  // ${className}RepositoryLocal({
+  //  required LocalDataService localDataService,
+  // }) : _localDataService = localDataService;
+
+  // final LocalDataService _localDataService;
+
+  // Add the implementation of the interface methods here
+}`;
+
+  const remoteRepositoryContent = `
+  import '${featureName}_repository.dart';
+/// Remote implementation of ${className}Repository
+class ${className}RepositoryRemote implements ${className}Repository {
+
+// In the constructor, you can add the dependencies needed to access the remote data
+// ${className}RepositoryRemote({
+//    required ApiClient apiClient,
+//  }) : _apiClient = apiClient;
+
+//  final ApiClient _apiClient;
+
+// Add the implementation of the interface methods here
+}`;
+
+  const repositoryFile = path.join(
+    rootPath,
+    `lib/data/repositories/${featureName}/${featureName}_repository.dart`
+  );
+
+  const localRepositoryFile = path.join(
+    rootPath,
+    `lib/data/repositories/${featureName}/${featureName}_repository_local.dart`
+  );
+
+  const remoteRepositoryFile = path.join(
+    rootPath,
+    `lib/data/repositories/${featureName}/${featureName}_repository_remote.dart`
+  );
+
+  await fs.promises.writeFile(repositoryFile, repositoryContent, { flag: "w" });
+  await fs.promises.writeFile(localRepositoryFile, localRepositoryContent, {
+    flag: "w",
+  });
+  await fs.promises.writeFile(remoteRepositoryFile, remoteRepositoryContent, {
+    flag: "w",
+  });
+}
+
+async function createFeatureInDomainLayer(
+  rootPath: string,
+  featureName: string
+) {
+  const featureFolders = [`lib/domain/models/${featureName}`];
+
+  for (const folder of featureFolders) {
+    const folderPath = path.join(rootPath, folder);
+    await fs.promises.mkdir(folderPath, { recursive: true });
+  }
+
+  // Create the code templates with the feature name
+  const className = capitalizeFeatureName(featureName);
+
+  const modelContent = `
+class ${className} {
+  /// Add the implementation of the model here
+}`;
+
+  const modelFile = path.join(
+    rootPath,
+    `lib/domain/models/${featureName}/${featureName}.dart`
+  );
+
+  await fs.promises.writeFile(modelFile, modelContent, { flag: "w" });
 }
 
 // This method is called when your extension is deactivated